backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Sparkstral backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.136.1",
    "mistralai>=2.4.2",
    "pydantic-settings>=2.14.0",
    "psycopg[binary]>=3.2.0",
    "sqlalchemy>=2.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
